Update package bc of giropay + paydirekt fusion
Important changes:
shoppingCartType is basically required, because if not provided it will be interpreted as MIXED type which requires shipping address. Now it sends shoppingCartType with default "ANONYMOUS_DONATION" because this type does not require additional shipping information, so existing calls to the wrapper shouldn't require any changes.

Also there is the field "kassenzeichen" wich can be used to display some additional data in Girocockpit.